[arcane,core] Utilise un 'Int16' au lieu d'un 'Int32' pour conserver le type de l'entité.
Le but est de diminuer la consommation mémoire lorsque le type de l'entité
sera conservé dans une variable.

+++ arcane/src/arcane/Item.h
@@ -177,7 +177,7 @@ class ARCANE_CORE_EXPORT Item
   Int32 owner() const { return m_internal->owner(); }
 
   //! Type de l'entité
-  Int32 type() const { return m_internal->typeId(); }
+  Int16 type() const { return m_internal->typeId(); }
 
   //! Type de l'entité
   ItemTypeId itemTypeId() const { return m_internal->itemTypeId(); }
